gmail_calendar_automation/requirements.txt
google-adk
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client